server/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
pydantic
python-dotenv
